requirements.txt
click==7.1.2
hydra-core==1.1.1
omegaconf==2.1.1
pytest==6.2.5
python-dotenv==0.19.2
setuptools==58.0.4
torch==1.10.1
torchvision==0.11.2
